[CORE-1136] Moved update code into a new file.
